[FIN-178] feat: JWT 적용 및 로그인 상태 유지 로직 수정
