+++ package.json
@@ -11,11 +11,13 @@
     "format:fix": "prettier --write ."
   },
   "dependencies": {
+    "@tailwindcss/typography": "^0.5.9",
     "@toast-ui/editor": "^3.2.2",
     "@types/node": "20.3.2",
     "@types/react": "18.2.14",
     "@types/react-dom": "18.2.6",
     "autoprefixer": "10.4.14",
+    "cookies-next": "^2.1.2",
     "eslint-config-next": "13.4.12",
     "next": "13.4.7",
     "postcss": "8.4.24",
@@ -26,14 +28,13 @@
     "remark-gfm": "^3.0.1",
     "tailwindcss": "3.3.2",
     "typescript": "5.1.6",
-    "zustand": "^4.3.9",
-    "@tailwindcss/typography": "^0.5.9"
+    "zustand": "^4.3.9"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.42.0",
     "@typescript-eslint/parser": "^5.42.0",
-    "eslint": "^8.45.0",
     "daisyui": "^3.5.0",
+    "eslint": "^8.45.0",
     "prettier": "^2.8.8"
   }
 }
